+++ bulbs/content/migrations/0010_content_instant_article_id.py
@@ -1,0 +1,19 @@
+# -*- coding: utf-8 -*-
+from __future__ import unicode_literals
+
+from django.db import migrations, models
+
+
+class Migration(migrations.Migration):
+
+    dependencies = [
+        ('content', '0009_auto_20160422_1212'),
+    ]
+
+    operations = [
+        migrations.AddField(
+            model_name='content',
+            name='instant_article_id',
+            field=models.IntegerField(default=None, null=True, blank=True),
+        ),
+    ]