+++ bulbs/contributions/csv_serializers.py
@@ -1,8 +1,9 @@
 from django.contrib.auth import get_user_model
+from django.utils import timezone
 
 from rest_framework import serializers
 
-from .models import Contribution
+from .models import Contribution, LineItem
 
 
 contributor_cls = get_user_model()
@@ -21,7 +22,7 @@ class ContributionCSVSerializer(serializers.ModelSerializer):
             'last_name': obj.contributor.last_name,
             'title': obj.content.title,
             'feature_type': obj.content.feature_type,
-            'publish_date': obj.content.published,
+            'publish_date': timezone.localtime(obj.content.published),
             'rate': obj.get_pay,
             'payroll_name': full_name
         }
@@ -30,4 +31,24 @@ class ContributionCSVSerializer(serializers.ModelSerializer):
             payroll_name = getattr(profile, 'payroll_name', None)
             if payroll_name:
                 data['payroll_name'] = payroll_name
+        return data
+
+
+class LineItemCSVSerializer(serializers.ModelSerializer):
+
+    class Meta:
+        model = LineItem
+
+    def to_representation(self, obj):
+        data = {
+            "amount": obj.amount,
+            "note": obj.note,
+            "payment_date": timezone.localtime(obj.payment_date),
+            "payroll_name": obj.contributor.get_full_name()
+        }
+        profile = getattr(obj.contributor, "freelanceprofile", None)
+        if profile:
+            payroll_name = getattr(profile, "payroll_name", None)
+            if payroll_name:
+                data["payroll_name"] = payroll_name
         return data