+++ bulbs/instant_articles/views.py
@@ -1,3 +1,4 @@
+from django.conf import settings
 from django.template import RequestContext, loader
 from django.template.base import TemplateDoesNotExist
 from django.views.decorators.cache import cache_control
@@ -5,6 +6,9 @@ from django.views.decorators.cache import cache_control
 from bulbs.content.models import Content
 from bulbs.content.views import BaseContentDetailView
 from bulbs.feeds.views import RSSView
+
+from bulbs.instant_articles.renderer import InstantArticleRenderer
+from bulbs.instant_articles.transform import transform
 
 
 class InstantArticleRSSView(RSSView):
@@ -28,9 +32,11 @@ class InstantArticleRSSView(RSSView):
 
         for content in context["page_obj"].object_list:
             content.feed_url = self.request.build_absolute_uri(content.get_absolute_url())
+            body = getattr(content, 'body', "")
             content_ctx = {
                 "content": content,
-                "absolute_uri": self.request.META.get('HTTP_HOST', None)
+                "absolute_uri": getattr(settings, 'WWW_URL'),
+                "transformed_body": transform(body, InstantArticleRenderer())
             }
             try:
                 content.instant_article_html = loader.render_to_string(
@@ -57,8 +63,10 @@ class InstantArticleContentView(BaseContentDetailView):
     def get_context_data(self, *args, **kwargs):
         context = super(InstantArticleContentView, self).get_context_data(*args, **kwargs)
         targeting = self.object.get_targeting()
+        body = getattr(self.object, 'body', "")
+        context["transformed_body"] = transform(body, InstantArticleRenderer())
         context["targeting"] = targeting
-        context["absolute_uri"] = self.request.META.get("HTTP_HOST", None)
+        context["absolute_uri"] = getattr(settings, "WWW_URL")
         return context