+++ tests/api/test_content_api.py
@@ -1,7 +1,6 @@
 import json
 from datetime import datetime, timedelta
 
-import elasticsearch
 from django.contrib.auth import get_user_model
 from django.conf.urls import url
 from django.core.urlresolvers import reverse
@@ -10,15 +9,21 @@ from django.test.client import Client
 from django.test.utils import override_settings
 from django.utils import timezone
 
+import elasticsearch
+
 import bulbs.api.urls
 from bulbs.content.models import LogEntry, Tag, Content, ObfuscatedUrlInfo
-from example.testcontent.models import TestContentObj, TestContentDetailImage
 from bulbs.utils.test import JsonEncoder, BaseAPITestCase, make_content
 
+from example.testcontent.models import TestContentObj, TestContentDetailImage
+
 try:
-    from unittest.mock import Mock
+    from unittest import mock
 except ImportError:
-    from mock import Mock
+    import mock
+
+
+User = get_user_model()
 
 
 class TestContentListingAPI(BaseAPITestCase):
@@ -308,13 +313,31 @@ class BaseUpdateContentAPI(BaseAPITestCase):
 class TestUpdateContentAPI(BaseUpdateContentAPI):
     """Tests updating an `Article`"""
 
+    def setUp(self):
+        super(TestUpdateContentAPI, self).setUp()
+        self.newguy = User.objects.create(
+            email="newguy@new.com",
+            username="tone"
+        )
+
     def create_content(self):
-        self.content = make_content(TestContentObj, title="Booyah: The Cramer Story", foo="booyah")
+        self.content = make_content(
+            TestContentObj, title="Booyah: The Cramer Story", foo="booyah", authors=[]
+        )
 
     def updated_data(self):
         return dict(
             title="Cramer 2: Electric Booyah-loo",
-            foo="whatta guy....booyah indeed!"
+            foo="whatta guy....booyah indeed!",
+            authors=[{
+                "last_name": "",
+                "username": self.newguy.username,
+                "first_name": "",
+                "email": self.newguy.email,
+                "short_name": "",
+                "id": self.newguy.id,
+                "full_name": ""
+            }]
         )
 
     def test_update_article(self):
@@ -349,7 +372,9 @@ class TestUpdateAuthorsAPI(BaseUpdateContentAPI):
         )
 
     def test_update_article(self):
-        self._test_update_content()
+        with mock.patch("django.core.mail.EmailMultiAlternatives.send") as mock_send:
+            self._test_update_content()
+            self.assertTrue(mock_send.called)
 
 
 class TestAddTagsAPI(BaseUpdateContentAPI):
@@ -590,15 +615,15 @@ class TestContentTypeSearchAPI(BaseAPITestCase):
         r = self.api_client.get(url, dict(search="two"), format="json")
         self.assertEqual(r.status_code, 200)
         self.assertEqual(len(r.data["results"]), 1)
-        # Content, TestContentObj, TestContentObjTwo, TestContentDetailImage,
-        # TestRecircContentObject
+        # Content, TestContentObj, TestContentObjTwo, TestContentObjThree,
+        # TestContentDetailImage, TestRecircContentObject
         r = self.api_client.get(url, dict(search="conte"), format="json")
         self.assertEqual(r.status_code, 200)
-        self.assertEqual(len(r.data["results"]), 5)
+        self.assertEqual(len(r.data["results"]), 6)
         # no query gives us all types
         r = self.api_client.get(url, format="json")
         self.assertEqual(r.status_code, 200)
-        self.assertEqual(len(r.data["results"]), 7)
+        self.assertEqual(len(r.data["results"]), 8)
 
 
 class TestContentResolveAPI(BaseAPITestCase):
@@ -622,7 +647,7 @@ class TestContentResolveAPI(BaseAPITestCase):
         # Simulate app-specific content URL
         class ContentUrls(object):
             urlpatterns = (bulbs.api.urls.urlpatterns +
-                           (url(r"^article/(?P<slug>[\w-]*)-(?P<pk>\d+)$", Mock()),))
+                           (url(r"^article/(?P<slug>[\w-]*)-(?P<pk>\d+)$", mock.Mock()),))
         make_content(id=123)
         with override_settings(ROOT_URLCONF=ContentUrls):
             r = self.resolve(url="/article/some-slug-123?one=two")
@@ -633,7 +658,7 @@ class TestContentResolveAPI(BaseAPITestCase):
         # Simulate app-specific content URL
         class ContentUrls(object):
             urlpatterns = (bulbs.api.urls.urlpatterns +
-                           (url(r"^article/(?P<slug>[\w-]*)$", Mock()),))
+                           (url(r"^article/(?P<slug>[\w-]*)$", mock.Mock()),))
         with override_settings(ROOT_URLCONF=ContentUrls):
             self.assertEqual(404, self.resolve(url="/article/some-slug").status_code)